chore(engine): Rename EntityQueueRequest to PlayerQueueRequest
